+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/ChatClient.kt
@@ -37,6 +37,7 @@ import io.getstream.chat.android.client.api.models.QueryThreadsRequest
 import io.getstream.chat.android.client.api.models.QueryUsersRequest
 import io.getstream.chat.android.client.api.models.SendActionRequest
 import io.getstream.chat.android.client.api.models.identifier.AddDeviceIdentifier
+import io.getstream.chat.android.client.api.models.identifier.ConnectUserIdentifier
 import io.getstream.chat.android.client.api.models.identifier.DeleteDeviceIdentifier
 import io.getstream.chat.android.client.api.models.identifier.DeleteMessageIdentifier
 import io.getstream.chat.android.client.api.models.identifier.DeleteReactionIdentifier
@@ -646,6 +647,7 @@ internal constructor(
             userScope.userId.value = user.id
             connectUserSuspend(user, tokenProvider, timeoutMilliseconds)
         }
+            .share(clientScope) { ConnectUserIdentifier(user) }
     }
 
     private suspend fun connectUserSuspend(
@@ -1507,6 +1509,14 @@ internal constructor(
                     Message(extraData = mapOf("poll_id" to poll.id)),
                 )
             }
+    }
+
+    @CheckResult
+    public fun suggestPollOption(
+        pollId: String,
+        option: String,
+    ): Call<Option> {
+        return api.suggestPollOption(pollId, option)
     }
 
     /**