+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/chat/android/ui/common/feature/messages/list/MessageListController.kt
@@ -2216,6 +2216,12 @@ public class MessageListController(
         }
     }
 
+    public fun addPollOption(poll: Poll, option: String) {
+        scope.launch {
+            chatClient.suggestPollOption(poll.id, option).await()
+        }
+    }
+
     /**
      * A class designed for error event propagation.
      *